% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_id}
\alias{check_id}
\title{for each dataframe in a list check if there is a column called id and subid}
\usage{
check_id(list_of_dfs)
}
\description{
for each dataframe in a list check if there is a column called id and subid
}
