% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_iffi_index}
\alias{create_iffi_index}
\title{create the index to join it later to the shapefile}
\usage{
create_iffi_index(list_of_dfs, log_idx)
}
\description{
create the index to join it later to the shapefile
}
