% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_additional_information.R
\name{get_additional_information}
\alias{get_additional_information}
\title{Get more information from the LPM DATA}
\usage{
get_additional_information(
  iffi10 = NULL,
 
    path_LPM_shape = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/Landslides/Iffi_db_xxxx_to_2018/exportperEurac2020/Shapefiles/LPM_Akten.shp",
 
    path_stato = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/Landslides/Iffi_db_xxxx_to_2018/exportperEurac2020/database/Stato delle conoscenze.xls",
  keep_lpm = c("ACCGE_IDEN", "GEOKAT_ID", "OGGETTO")
)
}
\arguments{
\item{path_LPM_shape}{Path to the LPM shapefile}

\item{keep_lpm}{Vector of columns from the lpm shapefile that we want to keep}

\item{path_IFFI10}{Path to the IFFI10 shapefile}

\item{dsdf}{}
}
\value{
an object of type \code{sf}
}
\description{
This function takes three paths and returns an object of type sf with all the columns from the LPM_shapefile
specified in the parameter \code{keep_lm}
}
