% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_information.R
\name{get_date_information}
\alias{get_date_information}
\title{function to information about the dates and the temporal information of landslides}
\usage{
get_date_information(df)
}
\arguments{
\item{df}{object as returned from \code{iffitoR::make_shapefile}}
}
\value{
a spatial dataframe of class \code{sf}

\itemize{
\item \code{date_info} indicating the available time information for that movement ("no date", "year", "month", "day")
\item \code{date} column of class \code{date} indicating the date (if available)
\item \code{year.int} year as integer
\item \code{month.int} month as integer
\item \code{day.int} day as integer
\item \code{year.posix} year (year-01-01) as object of class \code{date}
}
}
\description{
function to information about the dates and the temporal information of landslides
}
