% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_descriptions_shape}
\alias{join_descriptions_shape}
\title{join each table in a list of tables from the dict on the shape}
\usage{
join_descriptions_shape(list_of_dfs_dict, shape)
}
\description{
join each table in a list of tables from the dict on the shape
}
