% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_on_iffikodex}
\alias{join_on_iffikodex}
\title{joins a sinle sf-object with a column "PIFF_ID" to another table
with a column "iffi_kodex"}
\usage{
join_on_iffikodex(shape, table2)
}
\description{
joins a sinle sf-object with a column "PIFF_ID" to another table
with a column "iffi_kodex"
}
