% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{landsld}
\alias{landsld}
\title{Dataset with 10482 slides from South Tyrol and some additional informations (EPSG: 32362)}
\format{
An object of class \code{"sf"}
}
\usage{
data(landsld)
}
\description{
Dataset with 10482 slides from South Tyrol and some additional informations (EPSG: 32362)
}
\keyword{dataset}
