% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsld_ita.R
\docType{data}
\name{landsld_ita}
\alias{landsld_ita}
\title{Example landslide data from the iffi database with italian column-names (EPSG: 32632)}
\format{
An object of class \code{"sf"}
}
\usage{
data(landsld_ita)
}
\description{
Example landslide data from the iffi database with italian column-names (EPSG: 32632)
}
\keyword{dataset}
