% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_old.R
\docType{data}
\name{landsld_old}
\alias{landsld_old}
\title{Dataset with 8919 slides from South Tyrol and some additional informations (EPSG: 32362)}
\format{
An object of class \code{"sf"}
}
\usage{
data(landsld_old)
}
\description{
Dataset with 8919 slides from South Tyrol and some additional informations (EPSG: 32362)
}
\keyword{dataset}
