% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_list_tables}
\alias{make_list_tables}
\title{return a list of table names for a list of db connection}
\usage{
make_list_tables(conns)
}
\description{
return a list of table names for a list of db connection
}
