% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shapefile.R
\name{make_shapefile}
\alias{make_shapefile}
\title{The main funtion in order to retrieve the data from the iffi}
\usage{
make_shapefile(
  database_dir = NULL,
  attribute_database_name = NULL,
  dictionary_database_name = NULL,
  shapefile = NULL,
  attri = NULL,
  joins = NULL
)
}
\arguments{
\item{database_dir}{Path to the directory of the databases}

\item{attribute_database_name}{Name without extension of the attributes database}

\item{dictionary_database_name}{Name without extension of the dictionary database}

\item{shapefile}{A shape that cointains the iffi_kodex in a column called PIFF_ID}

\item{attri}{A vector of attributes to query directly from the attributes database}

\item{join}{A named list of the tables that need to be joined in order to get the desciptions from the dictionary}
}
\value{
An sf-oject
}
\description{
Generate a shapefile from a given spatial data-object and attributes as
produced by a query on the iffi database
}
\examples{
\dontrun{
database_dir = "data/database/",
attribute_database_name = "tbl_frane",
dictionary_database_name = "diz_frane",
shapefile = "data/Shapefiles/IFFI10_5.shp",
attri = c(
  "anno_min",
  "anno_max",
  "mese_min",
  "mese_max",
  "giorno_min",
  "giorno_max"
),
joins = list(
  "tbl_frane.Geologia.litologia" = c(
     "diz_frane.diz_litologie.litologia",
     "diz_frane.diz_litologie.nome_litologia"
  ),

  "tbl_frane.clas_ii_liv.movimento" = c(
     "diz_frane.diz_movimenti.movimento",
     "diz_frane.diz_movimenti.nome_movimento"
  ),
  "tbl_frane.Uso_Suolo.uso_suolo" = c(
     "diz_frane.diz_usi_suolo.uso_suolo",
     "diz_frane.diz_usi_suolo.nome_uso_suolo"
  )
),
plot = FALSE
)
}
}
