% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_vector_table_names}
\alias{make_vector_table_names}
\title{Creates a vector of table names from a RODBC connection}
\usage{
make_vector_table_names(db_conn)
}
\arguments{
\item{db_conn}{A object if type RODBC}
}
\value{
A single vector of table names for a single db connection
}
\description{
Creates a vector of table names from a RODBC connection
}
