% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_df_on_table_name}
\alias{return_df_on_table_name}
\title{return a specific dataframe from a named list of dataframes}
\usage{
return_df_on_table_name(list_of_dfs, table_name)
}
\description{
return a specific dataframe from a named list of dataframes
}
