% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{landsld}
\alias{landsld}
\title{Example landslide data from the iffi database}
\format{
An object of class \code{"sf"}
}
\usage{
data(landsld)
}
\description{
Example landslide data from the iffi database
}
\examples{
data(lndsld)
dim(lndsld)


}
\keyword{dataset}
