% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_connection}
\alias{set_connection}
\title{Functions that help to query the data from the databases and build the shapefile}
\usage{
set_connection(database_dir)
}
\description{
Functions that help to query the data from the databases and build the shapefile
}
